+++ src/redturtle/prenotazioni/tests/test_prenotazione_events.py
@@ -133,6 +133,42 @@ class TestSPrenotazioneEvents(unittest.TestCase):
             mail.get_payload()[0].get_payload(),
         )
 
+    def test_email_send_on_submit_and_confirm_if_not_autoconfirm(self):
+        self.folder_prenotazioni.notify_on_submit = True
+        self.folder_prenotazioni.notify_on_confirm = True
+        self.folder_prenotazioni.notify_on_submit_subject = (
+            f"{self.email_subject} on submit"
+        )
+        self.folder_prenotazioni.notify_on_submit_message = (
+            f"{self.email_message} on submit"
+        )
+        self.folder_prenotazioni.notify_on_confirm_subject = (
+            f"{self.email_subject} on confirm"
+        )
+        self.folder_prenotazioni.notify_on_confirm_message = (
+            f"{self.email_message} on confirm"
+        )
+
+        self.assertFalse(self.mailhost.messages)
+
+        booking = self.create_booking()
+
+        api.content.transition(booking, "confirm")
+
+        self.assertEqual(len(self.mailhost.messages), 2)
+
+        submit_mail = email.message_from_bytes(self.mailhost.messages[0])
+        confirm_mail = email.message_from_bytes(self.mailhost.messages[1])
+
+        self.assertIn(
+            f"{self.email_subject} on submit",
+            "".join([i for i in submit_mail.values()]),
+        )
+        self.assertIn(
+            f"{self.email_subject} on confirm",
+            "".join([i for i in confirm_mail.values()]),
+        )
+
     def test_email_send_on_reject(self):
         self.folder_prenotazioni.notify_on_refuse = True
         self.folder_prenotazioni.notify_on_refuse_subject = self.email_subject
@@ -159,17 +195,24 @@ class TestSPrenotazioneEvents(unittest.TestCase):
             mail.get_payload()[0].get_payload(),
         )
 
-    def test_email_not_sent_on_submit_if_on_confirm_true(self):
+    def test_email_not_sent_on_submit_if_on_confirm_true_and_autoconfirm(self):
+        self.folder_prenotazioni.auto_confirm = True
         self.folder_prenotazioni.notify_on_submit = True
         self.folder_prenotazioni.notify_on_confirm = True
-        self.folder_prenotazioni.notify_on_confirm_subject = self.email_subject
-        self.folder_prenotazioni.notify_on_confirm_message = self.email_message
+        self.folder_prenotazioni.notify_on_submit_subject = (
+            f"{self.email_subject} on submit"
+        )
+        self.folder_prenotazioni.notify_on_submit_message = (
+            f"{self.email_message} on submit"
+        )
+        self.folder_prenotazioni.notify_on_confirm_subject = (
+            f"{self.email_subject} on confirm"
+        )
+        self.folder_prenotazioni.notify_on_confirm_message = (
+            f"{self.email_message} on confirm"
+        )
 
-        self.assertFalse(self.mailhost.messages)
-
-        booking = self.create_booking()
-
-        api.content.transition(booking, "confirm")
+        self.create_booking()
 
         self.assertEqual(len(self.mailhost.messages), 1)
 
@@ -178,12 +221,40 @@ class TestSPrenotazioneEvents(unittest.TestCase):
         self.assertTrue(mail.is_multipart())
 
         self.assertIn(
-            self.email_subject,
+            f"{self.email_subject} on confirm",
             "".join([i for i in mail.values()]),
         )
+
+    def test_email_sent_on_submit_if_autoconfirm_and_not_on_confirm(self):
+        self.folder_prenotazioni.notify_on_submit = True
+        self.folder_prenotazioni.notify_on_confirm = False
+        self.folder_prenotazioni.notify_on_submit_subject = (
+            f"{self.email_subject} on submit"
+        )
+        self.folder_prenotazioni.notify_on_submit_message = (
+            f"{self.email_message} on submit"
+        )
+        self.folder_prenotazioni.notify_on_confirm_subject = (
+            f"{self.email_subject} on confirm"
+        )
+        self.folder_prenotazioni.notify_on_confirm_message = (
+            f"{self.email_message} on confirm"
+        )
+        self.folder_prenotazioni.auto_confirm = True
+
+        self.assertFalse(self.mailhost.messages)
+
+        self.create_booking()
+
+        self.assertEqual(len(self.mailhost.messages), 1)
+
+        mail = email.message_from_bytes(self.mailhost.messages[0])
+
+        self.assertTrue(mail.is_multipart())
+
         self.assertIn(
-            self.email_message,
-            mail.get_payload()[0].get_payload(),
+            f"{self.email_subject} on submit",
+            "".join([i for i in mail.values()]),
         )
 
     def test_email_send_on_prenotazione_move(self):
@@ -209,3 +280,118 @@ class TestSPrenotazioneEvents(unittest.TestCase):
             self.email_message,
             mail.get_payload()[0].get_payload(),
         )
+
+    def test_email_send_on_prenotazione_move_has_ical(self):
+        self.folder_prenotazioni.notify_on_move = True
+        self.folder_prenotazioni.notify_on_move_subject = self.email_subject
+        self.folder_prenotazioni.notify_on_move_message = self.email_message
+
+        self.assertFalse(self.mailhost.messages)
+        booking = self.create_booking()
+        notify(MovedPrenotazione(booking))
+
+        self.assertEqual(len(self.mailhost.messages), 1)
+
+        mail_sent = self.mailhost.messages[0]
+        message = email.message_from_bytes(mail_sent)
+
+        self.assertTrue(len(message.get_payload()), 2)
+
+        attachment = message.get_payload()[1]
+        data = attachment.get_payload()
+
+        self.assertTrue(message.is_multipart())
+        self.assertEqual(attachment.get_filename(), "foo.ics")
+        self.assertIn("SUMMARY:Booking for Prenota foo", data)
+        if api.env.plone_version() < "6":
+            self.assertIn(
+                f'DTSTART;VALUE=DATE-TIME:{booking.booking_date.strftime("%Y%m%dT%H%M%S")}',
+                data,
+            )
+            self.assertIn(
+                f'DTEND;VALUE=DATE-TIME:{booking.booking_expiration_date.strftime("%Y%m%dT%H%M%S")}',
+                data,
+            )
+        else:
+            self.assertIn(
+                f'DTSTART:{booking.booking_date.strftime("%Y%m%dT%H%M%S")}',
+                data,
+            )
+            self.assertIn(
+                f'DTEND:{booking.booking_expiration_date.strftime("%Y%m%dT%H%M%S")}',
+                data,
+            )
+        self.assertIn(f"UID:{booking.UID()}", data)
+
+    def test_email_send_on_submit_does_not_have_ical(self):
+        self.folder_prenotazioni.notify_on_submit = True
+        self.folder_prenotazioni.notify_on_submit_subject = self.email_subject
+        self.folder_prenotazioni.notify_on_submit_message = self.email_message
+
+        self.assertFalse(self.mailhost.messages)
+        self.create_booking()
+
+        self.assertEqual(len(self.mailhost.messages), 1)
+
+        mail_sent = self.mailhost.messages[0]
+        message = email.message_from_bytes(mail_sent)
+
+        self.assertTrue(len(message.get_payload()), 1)
+
+    def test_email_send_on_confirm_has_ical(self):
+        self.folder_prenotazioni.notify_on_confirm = True
+        self.folder_prenotazioni.notify_on_confirm_subject = self.email_subject
+        self.folder_prenotazioni.notify_on_confirm_message = self.email_message
+        self.folder_prenotazioni.auto_confirm = True
+
+        self.assertFalse(self.mailhost.messages)
+        booking = self.create_booking()
+
+        self.assertEqual(len(self.mailhost.messages), 1)
+
+        mail_sent = self.mailhost.messages[0]
+        message = email.message_from_bytes(mail_sent)
+
+        self.assertTrue(len(message.get_payload()), 2)
+
+        attachment = message.get_payload()[1]
+        data = attachment.get_payload()
+
+        self.assertTrue(message.is_multipart())
+        self.assertEqual(attachment.get_filename(), "foo.ics")
+        self.assertIn("SUMMARY:Booking for Prenota foo", data)
+        if api.env.plone_version() < "6":
+            self.assertIn(
+                f'DTSTART;VALUE=DATE-TIME:{booking.booking_date.strftime("%Y%m%dT%H%M%S")}',
+                data,
+            )
+            self.assertIn(
+                f'DTEND;VALUE=DATE-TIME:{booking.booking_expiration_date.strftime("%Y%m%dT%H%M%S")}',
+                data,
+            )
+        else:
+            self.assertIn(
+                f'DTSTART:{booking.booking_date.strftime("%Y%m%dT%H%M%S")}',
+                data,
+            )
+            self.assertIn(
+                f'DTEND:{booking.booking_expiration_date.strftime("%Y%m%dT%H%M%S")}',
+                data,
+            )
+        self.assertIn(f"UID:{booking.UID()}", data)
+
+    def test_email_send_on_refuse_does_not_have_ical(self):
+        self.folder_prenotazioni.notify_on_refuse = True
+        self.folder_prenotazioni.notify_on_refuse_subject = self.email_subject
+        self.folder_prenotazioni.notify_on_refuse_message = self.email_message
+
+        self.assertFalse(self.mailhost.messages)
+        booking = self.create_booking()
+        api.content.transition(booking, "refuse")
+
+        self.assertEqual(len(self.mailhost.messages), 1)
+
+        mail_sent = self.mailhost.messages[0]
+        message = email.message_from_bytes(mail_sent)
+
+        self.assertTrue(len(message.get_payload()), 1)